// src/lib/types.ts
export interface VulnmapProject {
  name: string;
  id: string;
  created: string;
  origin: string;
  type: string;
  branch: string | null;
  status: 'inactive' | 'active';
}